% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{check_id_vars_col}
\alias{check_id_vars_col}
\title{Check variable col in df for factor type, and coerce to character}
\usage{
check_id_vars_col(df, col)
}
\arguments{
\item{df}{input data.frame}

\item{col}{column name to check}
}
\value{
\code{df}, without factor column on id var col
}
\description{
Check variable col in df for factor type, and coerce to character
}
