% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-set-retrieval-scores_helpers.r
\name{check_repair_relevance_compare}
\alias{check_repair_relevance_compare}
\title{Internal helper function to check comparison matrix for
inconsistent gold_standard and relevance values}
\usage{
check_repair_relevance_compare(gold_vs_pred)
}
\arguments{
\item{gold_vs_pred}{as created by \code{create_comparison}}
}
\value{
comparison matrix compatible with compute_intermediate_results
}
\description{
Internal helper function to check comparison matrix for
inconsistent gold_standard and relevance values
}
