% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-set-retrieval-scores_helpers.r
\name{check_repair_relevance_predicted}
\alias{check_repair_relevance_predicted}
\title{Internal helper to check predicted data frame for valid relevance column}
\usage{
check_repair_relevance_predicted(predicted)
}
\arguments{
\item{predicted}{expects data.frame with column \emph{"relevance"}}
}
\value{
valid predicted data.frame with possibly eliminated missing values
}
\description{
Internal helper to check predicted data frame for valid relevance column
}
