% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank_col.R
\name{create_rank_col}
\alias{create_rank_col}
\alias{create_rank_col_dplyr}
\title{Create a rank per doc_id based on score}
\usage{
create_rank_col(df)

create_rank_col_dplyr(df)
}
\arguments{
\item{df}{with columns "score" and "doc_id"}
}
\value{
\code{df} with extra column "rank"
}
\description{
Create a rank per doc_id based on score
}
\section{Functions}{
\itemize{
\item \code{create_rank_col_dplyr()}: variant with internal usage of
dplyr rather than collapse library.

}}
