% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnb_gold_standard}
\alias{dnb_gold_standard}
\title{DNB gold standard data for computing evaluation metrics}
\format{
\subsection{\code{dnb_gold_standard}}{

A data frame with 337 rows and 2 columns:
\describe{
\item{doc_id}{DNB identifier of a document in catalogue}
\item{label_id}{DNB identifier of a concept in the GND subject vocabulary}
}
}
}
\usage{
dnb_gold_standard
}
\description{
A subset of documents found in the catalogue of the DNB with intelectually
assigned subject labels from the GND subject vocabulary.
The document ids match those in the \code{dnb_test_predictions} dataset.
}
\keyword{datasets}
