% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnb_label_distribution}
\alias{dnb_label_distribution}
\title{DNB label distribution for computing propensity scored metrics}
\format{
\subsection{\code{dnb_test_predictions}}{

A data frame with 100,000 rows and 4 columns:
\describe{
\item{label_id}{DNB identifier of a concept in the GND subject vocabulary}
\item{label_freq}{Number of occurences of the specified label in the overall catalogue}
\item{n_docs}{Overall number of documents in the ground truth dataset}
}
}
}
\usage{
dnb_label_distribution
}
\description{
A subset of labels used in the catalogue of the dnb along with their
frequencies of occurence. The label_ids match those in the
\code{dnb_gold_standard} and \code{dnb_test_predictions} datasets.
}
\keyword{datasets}
