% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnb_test_predictions}
\alias{dnb_test_predictions}
\title{DNB test predictions for computing evaluation metrics}
\format{
\subsection{\code{dnb_test_predictions}}{

A data frame with 100,000 rows and 4 columns:
\describe{
\item{doc_id}{DNB identifier of a document in catalogue}
\item{label_id}{DNB identifier of a concept in the GND subject vocabulary}
\item{score}{a confidence score generated by the indexing method}
\item{rank}{ranks based on the score-columns}
}
}
}
\usage{
dnb_test_predictions
}
\description{
A subset of documents found in the catalogue of the DNB with predictions
generated with some arbitrary indexing method. The document ids match those
in the \code{dnb_gold_standard} dataset.
}
\keyword{datasets}
