% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_helpers.r
\name{helper_f}
\alias{helper_f}
\title{wrapper for compute_intermediate and summarise on one bt_sample}
\usage{
helper_f(
  sampled_id_list,
  compare_cpy,
  grouping_var,
  label_distribution = NULL,
  ps_flags = list(intermed = FALSE, summarise = FALSE),
  cost_fp = NULL,
  replace_zero_division_with = options::opt("replace_zero_division_with"),
  drop_empty_groups = options::opt("drop_empty_groups")
)
}
\arguments{
\item{sampled_id_list}{doc_ids of this bootstrap}

\item{compare_cpy}{as created by \code{create_comparison}}

\item{grouping_var}{variables for aggreation}

\item{label_distribution}{as in compute_set_retrieval_scores}

\item{ps_flags}{list as returned by \code{set_ps_flags}}

\item{cost_fp}{numeric > 0}

\item{replace_zero_division_with}{In macro averaged results (doc-avg, subj-avg) it may occur that some
instances have no predictions or no gold-standard. In these cases calculating
precision and recall may lead to zero divisions. CASIMiR standardly removes
these missing values from macro averages, leading to a smaller \code{support}
(count of instances that were averaged). Other implementations of macro
averaged precision and recall default to 0 in these cases. This option
will allow to control the default. Set any value between 0 and 1. (Defaults to \code{NULL}, overwritable using option 'casimir.replace_zero_division_with' or environment variable 'R_CASIMIR_REPLACE_ZERO_DIVISION_WITH')}

\item{drop_empty_groups}{Should empty levels of factor variables be droped in grouped set retrieval
computation? (Defaults to \code{TRUE}, overwritable using option 'casimir.drop_empty_groups' or environment variable 'R_CASIMIR_DROP_EMPTY_GROUPS')}
}
\value{
as \code{summarise_intermediate_results}
}
\description{
wrapper for compute_intermediate and summarise on one bt_sample
}
