% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-set-retrieval-scores_helpers.r
\name{helper_f_dplyr_version}
\alias{helper_f_dplyr_version}
\title{Internal wrapper for compute_intermediat and summarise on one bt_sample}
\usage{
helper_f_dplyr_version(
  sampled_id_list,
  compare_cpy,
  grouping_var,
  label_distribution = NULL,
  ps_flags = list(intermed = FALSE, summarise = FALSE),
  cost_fp = NULL
)
}
\arguments{
\item{sampled_id_list}{doc_ids of this bootstrap}

\item{compare_cpy}{as created by \code{create_comparison}}

\item{grouping_var}{variables for aggreation}

\item{label_distribution}{as in compute_set_retrieval_scores}

\item{ps_flags}{list with logicals "intermed" and "summarise"}

\item{cost_fp}{numeric > 0}
}
\value{
as \code{summarise_intermediate_results_dplyr_version}
}
\description{
Internal wrapper for compute_intermediat and summarise on one bt_sample
}
