% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_retrieval_scores_reference.r
\name{lrap_score}
\alias{lrap_score}
\title{Reference implementation for Label Ranking Average Precision}
\usage{
lrap_score(gold_vs_pred)
}
\arguments{
\item{gold_vs_pred}{a \code{data.frame} as generated by create_comparison}
}
\value{
numeric value of lrap
}
\description{
Reference implementation for Label Ranking Average Precision
}
