% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_retrieval_scores_reference.r
\name{ndcg_score}
\alias{ndcg_score}
\title{Helper function for document-wise computation of ranked retrieval scores}
\usage{
ndcg_score(gold_vs_pred, limit = NULL)
}
\arguments{
\item{gold_vs_pred}{A data.frame as generated by \code{create_comparison}.}

\item{limit}{An integer cutoff value for NDCG@N.}
}
\value{
The numeric value of NDCG.
}
\description{
Helper function for document-wise computation of ranked retrieval scores
DCG, NDCG and LRAP.
Implemented as in Annif
\url{https://github.com/NatLibFi/Annif/blob/master/annif/eval.py}.
Reference implementation for NDCG to test against.
}
