% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casimir_package.R
\name{options}
\alias{options}
\title{casimir Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{casimir} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "casimir")

options::opt(x, default, env = "casimir")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{ignore_inconsistencies}{\describe{
Warnings about data inconsistencies will be silenced.\item{default: }{\preformatted{FALSE}}
\item{option: }{casimir.ignore_inconsistencies}
\item{envvar: }{R_CASIMIR_IGNORE_INCONSISTENCIES (evaluated if possible, raw string otherwise)}
}}

\item{progress}{\describe{
Display progress bars for iterated computations (like bootstrap CI or
PR-Curves)\item{default: }{\preformatted{FALSE}}
\item{option: }{casimir.progress}
\item{envvar: }{R_CASIMIR_PROGRESS (evaluated if possible, raw string otherwise)}
}}

\item{verbose}{\describe{
Verbose reorting of computation steps for debugging\item{default: }{\preformatted{FALSE}}
\item{option: }{casimir.verbose}
\item{envvar: }{R_CASIMIR_VERBOSE (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.setenv Sys.getenv
}
