% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_curve.r
\name{pr_curve_post_processing}
\alias{pr_curve_post_processing}
\title{Reshape pr_curve_data to a format that is easier for plotting}
\usage{
pr_curve_post_processing(results_summary)
}
\arguments{
\item{results_summary}{as produced by summarise_intermediate_results}
}
\value{
a \code{data.frame} with cols \code{c("searchspace_id", "prec", "rec", "prec_cummax")}
and possibly additional stratification variables
}
\description{
Reshape pr_curve_data to a format that is easier for plotting
}
