% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{process_cost_fp}
\alias{process_cost_fp}
\title{Process input for cost_fp}
\usage{
process_cost_fp(cost_fp_constant, gold_vs_pred)
}
\arguments{
\item{cost_fp_constant}{Constant cost assigned to false positives.
cost_fp_constant must be
a numeric value > 0 or one of 'max', 'min', 'mean' (computed with reference
to the \code{gold_standard} label distribution). The default is NULL, i.e.
label weights are applied to false positives as to false negatives and
true positives.}

\item{gold_vs_pred}{\code{data.frame} with logical columns
\emph{"suggested", "gold"} as produced by \code{create_comparison}}
}
\value{
a numeric value > 0
}
\description{
Process input for cost_fp
}
