% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_intermediate_results.r
\name{summarise_intermediate_results_dplyr}
\alias{summarise_intermediate_results_dplyr}
\title{Compute mean of intermediate results. Variant with dplyr based
internals rather then collapse internals}
\usage{
summarise_intermediate_results_dplyr(
  intermediate_results,
  propensity_scored = FALSE,
  label_distribution = NULL
)
}
\arguments{
\item{intermediate_results}{with cols "prec", "rprec", "rec", "f1" as
produced by compute intermediate results}

\item{propensity_scored}{as in \code{compute_set_retrieval_scores}}

\item{label_distribution}{as in \code{compute_set_retrieval_scores}}
}
\value{
data.frame with cols metric, value
}
\description{
Compute mean of intermediate results. Variant with dplyr based
internals rather then collapse internals
}
