% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_auc.r
\name{boot_worker_fn}
\alias{boot_worker_fn}
\title{A wrapper for use within bootstrap computation of auc
which covers the repeated application of
\enumerate{
\item join with resampled doc_ids
\item summarise_intermediate_results
\item post_processing_of_curve data
\item auc computation
}}
\usage{
boot_worker_fn(
  sampled_id_list,
  intermed_res,
  propensity_scored,
  replace_zero_division_with
)
}
\arguments{
\item{sampled_id_list}{doc_ids of the examples drawn in each bootstrap
iteration}

\item{intermed_res}{intermediate results as produced by
\code{compute_intermediate_results}, with a column \code{"searchspace_id"}
as grouping variable}

\item{propensity_scored}{logical, whether to use propensity scores as weights}

\item{replace_zero_division_with}{In macro averaged results (doc-avg, subj-avg) it may occur that some
instances have no predictions or no gold-standard. In these cases calculating
precision and recall may lead to zero divisions. CASIMiR standardly removes
these missing values from macro averages, leading to a smaller \code{support}
(count of instances that were averaged). Other implementations of macro
averaged precision and recall default to 0 in these cases. This option
will allow to control the default. Set any value between 0 and 1. (Defaults to \code{NULL}, overwritable using option 'casimir.replace_zero_division_with' or environment variable 'R_CASIMIR_REPLACE_ZERO_DIVISION_WITH')}
}
\value{
a \code{data.frame} with col pr_auc and potential grouping_vars
}
\description{
A wrapper for use within bootstrap computation of auc
which covers the repeated application of
\enumerate{
\item join with resampled doc_ids
\item summarise_intermediate_results
\item post_processing_of_curve data
\item auc computation
}
}
