% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_auc.r
\name{boot_worker_fn}
\alias{boot_worker_fn}
\title{A wrapper for use within bootstrap computation of auc
which covers the repeated application of
\enumerate{
\item join with resampled doc_ids
\item summarise_intermediate_results
\item post_processing_of_curve data
\item auc computation
}}
\usage{
boot_worker_fn(sampled_id_list, intermed_res, propensity_scored)
}
\arguments{
\item{sampled_id_list}{doc_ids of the examples drawn in each bootstrap
iteration}

\item{intermed_res}{intermediate results as produced by
\code{compute_intermediate_results}, with a column \code{"searchspace_id"}
as grouping variable}

\item{propensity_scored}{as in \code{compute_pr_auc}}
}
\value{
a \code{data.frame} with col pr_auc and potential grouping_vars
}
\description{
A wrapper for use within bootstrap computation of auc
which covers the repeated application of
\enumerate{
\item join with resampled doc_ids
\item summarise_intermediate_results
\item post_processing_of_curve data
\item auc computation
}
}
