% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{check_id_vars}
\alias{check_id_vars}
\title{Internal helper function designed to ensure that id columns are not
passed as factor variables. Factor variables in id columns may
cause undesired behaviour with the drop_emtpy_group argument}
\usage{
check_id_vars(df)
}
\arguments{
\item{df}{data.frame to check id columns}
}
\value{
\code{df}, without factor column on id vars
}
\description{
Internal helper function designed to ensure that id columns are not
passed as factor variables. Factor variables in id columns may
cause undesired behaviour with the drop_emtpy_group argument
}
