% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{check_id_vars_col}
\alias{check_id_vars_col}
\title{Coerce column to character}
\usage{
check_id_vars_col(df, col)
}
\arguments{
\item{df}{An input data.frame.}

\item{col}{The name of the column to check.}
}
\value{
The input data.frame \code{df} with the specified column being no
longer a factor variable.
}
\description{
Check an arbitrary column in a data.frame for factor type and coerce to
character.
}
