% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repair_relevance.r
\name{check_repair_relevance_compare}
\alias{check_repair_relevance_compare}
\title{Check for inconsistent relevance values}
\usage{
check_repair_relevance_compare(
  gold_vs_pred,
  ignore_inconsistencies = options::opt("ignore_inconsistencies")
)
}
\arguments{
\item{gold_vs_pred}{As created by \code{create_comparison}.}

\item{ignore_inconsistencies}{Warnings about data inconsistencies will be silenced. (Defaults to \code{FALSE}, overwritable using option 'casimir.ignore_inconsistencies' or environment variable 'R_CASIMIR_IGNORE_INCONSISTENCIES')}
}
\value{
A valid comparison matrix with possibly corrected relevance values,
being compatible with \code{compute_intermediate_results}.
}
\description{
Internal helper function to check a comparison matrix for inconsistent
relevance values of gold standard and predicted labels.
}
