% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_repair_relevance.r
\name{check_repair_relevance_pred}
\alias{check_repair_relevance_pred}
\title{Internal helper to check predicted data frame for valid relevance column}
\usage{
check_repair_relevance_pred(predicted)
}
\arguments{
\item{predicted}{multi-label prediction results. expects \code{data.frame}
with cols \emph{"label_id", "doc_id", "relevance"}}
}
\value{
valid predicted data.frame with possibly eliminated missing values
}
\description{
Internal helper to check predicted data frame for valid relevance column
}
