% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_intermediate_results.r
\name{compute_intermediate_results}
\alias{compute_intermediate_results}
\alias{compute_intermediate_results_dplyr}
\title{Compute true positives, false positives on aggregation level specified}
\usage{
compute_intermediate_results(
  gold_vs_pred,
  grouping_var,
  propensity_scored = FALSE,
  cost_fp = NULL,
  check_group_names = options::opt("check_group_names")
)

compute_intermediate_results_dplyr(
  gold_vs_pred,
  grouping_var,
  propensity_scored = FALSE,
  cost_fp = NULL
)
}
\arguments{
\item{gold_vs_pred}{\code{data.frame} with logical columns
\emph{"suggested", "gold"} as produced by \code{create_comparison}}

\item{grouping_var}{a character vector of variables that must be present in
gold_vs_pred (dplyr version requires rlang symbols)}

\item{propensity_scored}{logical, whether to use propensity scores as weights}

\item{cost_fp}{numeric > 0, default is NULL}

\item{check_group_names}{Perform replacement of dots in grouping columns. Disable for faster
computation, if you can make sure that all columns used for grouping
(doc_id, label_id, doc_groups, label_groups) do not contain dots (Defaults to \code{TRUE}, overwritable using option 'casimir.check_group_names' or environment variable 'R_CASIMIR_CHECK_GROUP_NAMES')}
}
\value{
data.frame with cols "n_gold", "n_suggested", "tp", "fp", "fn",
"prec", "rec", "f1"
}
\description{
Compute true positives, false positives on aggregation level specified
}
\section{Functions}{
\itemize{
\item \code{compute_intermediate_results_dplyr()}: variant with dplyr based
internals rather then collapse internals

}}
\examples{

gold <- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "b",
  "A", "c",
  "B", "a",
  "B", "d",
  "C", "a",
  "C", "b",
  "C", "d",
  "C", "f"
)

pred <- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "d",
  "A", "f",
  "B", "a",
  "B", "e",
  "C", "f"
)

gold_vs_pred <- create_comparison(gold, pred)
casimir:::compute_intermediate_results(gold_vs_pred, "doc_id")

}
