% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-set-retrieval-scores_helpers.r
\name{compute_intermediate_results}
\alias{compute_intermediate_results}
\alias{compute_intermediate_results_dplyr_version}
\title{Compute true positives, false positives on aggregation level specified}
\usage{
compute_intermediate_results(
  gold_vs_pred,
  grouping_var,
  propensity_scored = FALSE,
  cost_fp = NULL
)

compute_intermediate_results_dplyr_version(
  gold_vs_pred,
  grouping_var,
  propensity_scored = FALSE,
  cost_fp = NULL
)
}
\arguments{
\item{gold_vs_pred}{\code{data.frame} with logical columns
\emph{"suggested", "gold"} as produced by \code{create_comparison}}

\item{grouping_var}{a character vector of variables that must be present in
gold_vs_pred (dplyr version requires rlang symbols)}

\item{propensity_scored}{logical, whether to use propensity scores as weights}

\item{cost_fp}{numeric > 0, default is NULL}
}
\value{
data.frame with cols "n_gold", "n_suggested", "tp", "fp", "fn", "prec", "rec", "f1"
}
\description{
Compute true positives, false positives on aggregation level specified
}
\section{Functions}{
\itemize{
\item \code{compute_intermediate_results_dplyr_version()}: variant with dplyr based
internals rather then collapse internals

}}
\examples{

gold <- tibble::tribble(
    ~doc_id, ~label_id,
    "A", "a",
    "A", "b",
    "A", "c",
    "B", "a",
    "B", "d",
    "C", "a",
    "C", "b",
    "C", "d",
    "C", "f"
)

pred<- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "d",
  "A", "f",
  "B", "a",
  "B", "e",
  "C", "f"
)

gold_vs_pred <- create_comparison(gold, pred)
casimir:::compute_intermediate_results(gold_vs_pred, "doc_id")

}
