% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_intermedtiate_results_rr.r
\name{compute_intermediate_results_rr}
\alias{compute_intermediate_results_rr}
\title{Compute intermediate ranked retrieval results}
\usage{
compute_intermediate_results_rr(
  gold_vs_pred,
  grouping_var,
  drop_empty_groups = options::opt("drop_empty_groups")
)
}
\arguments{
\item{gold_vs_pred}{a \code{data.frame} as generated by create_comparison
a column \emph{"score"} is expected}

\item{grouping_var}{character vector of variables that must be present in
base_compare}

\item{drop_empty_groups}{Should empty levels of factor variables be droped in grouped set retrieval
computation? (Defaults to \code{TRUE}, overwritable using option 'casimir.drop_empty_groups' or environment variable 'R_CASIMIR_DROP_EMPTY_GROUPS')}
}
\value{
a \code{data.frame} with cols \emph{"dcg", "idcg", "ndcg", "lrap"}
}
\description{
Compute intermediate ranked retrieval results
}
\examples{

gold <- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "b",
  "A", "c",
  "A", "d",
  "A", "e",
)

pred <- tibble::tribble(
  ~doc_id, ~label_id, ~score,
  "A", "f", 0.3277,
  "A", "e", 0.32172,
  "A", "b", 0.13517,
  "A", "g", 0.10134,
  "A", "h", 0.09152,
  "A", "a", 0.07483,
  "A", "i", 0.03649,
  "A", "j", 0.03551,
  "A", "k", 0.03397,
  "A", "c", 0.03364
)

gold_vs_pred <- create_comparison(gold, pred)

casimir:::compute_intermediate_results_rr(
  gold_vs_pred,
  rlang::syms(c("doc_id"))
)
}
