% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank_col.R
\name{create_rank_col}
\alias{create_rank_col}
\alias{create_rank_col_dplyr}
\title{Create a rank column}
\usage{
create_rank_col(df)

create_rank_col_dplyr(df)
}
\arguments{
\item{df}{A data.frame with columns \code{"doc_id", "score"}.}
}
\value{
The input data.frame \code{df} with an additional column
\code{"rank"}.
}
\description{
Create a rank per document id based on score.
}
\section{Functions}{
\itemize{
\item \code{create_rank_col_dplyr()}: Variant with internal usage of
dplyr rather than collapse library.

}}
