% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_retrieval_scores_reference.r
\name{dcg_score}
\alias{dcg_score}
\title{Helper Function for document-wise computation of ranked retrieval scores
dcg, ndcg and lrap}
\usage{
dcg_score(df, limit = NULL)
}
\arguments{
\item{df}{a \code{data.frame} as generated by create_comparison}

\item{limit}{cut off value for dcgAtN}
}
\description{
Implemented as in Annif
https://github.com/NatLibFi/Annif/blob/master/annif/eval.py
Reference implementation of dcg to test against
}
