% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ranked-retrieval-scores_helpers.r
\name{dcg_score}
\alias{dcg_score}
\title{Reference implementation of dcg to test against}
\usage{
dcg_score(df, limit = NULL)
}
\arguments{
\item{df}{a \code{data.frame} as generated by create_comparison}

\item{limit}{cut off value for dcgAtN}
}
\description{
Reference implementation of dcg to test against
}
