% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_retrieval_scores_reference.r
\name{dcg_score}
\alias{dcg_score}
\title{Helper function for document-wise computation of ranked retrieval scores}
\usage{
dcg_score(df, limit = NULL)
}
\arguments{
\item{limit}{An integer cutoff value for DCG@N.}

\item{gold_vs_pred}{A data.frame as generated by \code{create_comparison}.}
}
\value{
The numeric value of DCG.
}
\description{
Helper function for document-wise computation of ranked retrieval scores
DCG, NDCG and LRAP.
Implemented as in Annif
\url{https://github.com/NatLibFi/Annif/blob/master/annif/eval.py}.
Reference implementation of DCG to test against.
}
