% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnb_label_distribution}
\alias{dnb_label_distribution}
\title{DNB label distribution for computing propensity scored metrics}
\format{
\subsection{\code{dnb_label_distribution}}{

A data frame with 7,772 rows and 3 columns:
\describe{
\item{\code{label_id}}{DNB identifier of a concept in the GND subject
vocabulary.}
\item{\code{label_freq}}{Number of occurences of the specified label in the
overall catalogue.}
\item{\code{n_docs}}{Overall number of documents in the ground truth
dataset.}
}
}
}
\usage{
dnb_label_distribution
}
\description{
A subset of labels used in the catalogue of the DNB along with their
frequencies of occurrence. The label_ids match those in the
\code{dnb_gold_standard} and \code{dnb_test_predictions} datasets.
}
\keyword{datasets}
