% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnb_test_predictions}
\alias{dnb_test_predictions}
\title{DNB test predictions for computing evaluation metrics}
\format{
\subsection{\code{dnb_test_predictions}}{

A data frame with 100,000 rows and 3 columns:
\describe{
\item{\code{doc_id}}{DNB identifier of a document in the catalogue.}
\item{\code{label_id}}{DNB identifier of a concept in the GND subject
vocabulary.}
\item{\code{score}}{A confidence score in \eqn{[0, 1]} generated by the
indexing method.}
}
}
}
\usage{
dnb_test_predictions
}
\description{
A subset of documents found in the catalogue of the DNB with predictions
generated with some arbitrary indexing method. The document ids match those
in the \code{dnb_gold_standard} dataset.
}
\keyword{datasets}
