% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ps_rprec_deno.r
\name{find_ps_rprec_deno}
\alias{find_ps_rprec_deno}
\alias{find_ps_rprec_deno_dplyr}
\title{Compute the denominator for rprecision, based on propensity scored
ranking of gold_standrad labels}
\usage{
find_ps_rprec_deno(gold_vs_pred, grouping_var, cost_fp)

find_ps_rprec_deno_dplyr(gold_vs_pred, grouping_var, cost_fp)
}
\arguments{
\item{gold_vs_pred}{data.frame with colums "score.pred", "score.gold"}

\item{grouping_var}{a character vector of variables that must be present in
gold_vs_pred (dplyr version requires rlang symbols)}

\item{cost_fp}{numeric > 0, default is NULL}
}
\value{
data.frame with cols "n_gold", "n_suggested", "tp", "fp",
"fn", "rprec_deno"
}
\description{
Compute the denominator for rprecision, based on propensity scored
ranking of gold_standrad labels
}
\section{Functions}{
\itemize{
\item \code{find_ps_rprec_deno_dplyr()}: variant with dplyr based
internals rather then collapse internals

}}
