% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pr_curve.r
\name{generate_pr_auc_replica}
\alias{generate_pr_auc_replica}
\title{Compute bootstrap replica of pr-auc}
\usage{
generate_pr_auc_replica(
  intermed_res_all_thrsld,
  seed,
  n_bt,
  propensity_scored,
  .progress
)
}
\arguments{
\item{intermed_res_all_thrsld}{intermediate results as produced by
\code{compute_intermediate_results}, with a column \code{"searchspace_id"}
as grouping variable}

\item{seed}{set seed for reproducibility of bootstrap run}

\item{n_bt}{integer number of bootstrap-replica to draw}

\item{propensity_scored}{as in \code{compute_pr_auc}}

\item{.progress}{show progress bar}
}
\value{
\code{data.frame} with cols \code{c("boot_replicate", "pr_auc")}
}
\description{
Helper function whcih performs the major bootstrap operation and wraps the
repeated application of \code{summarise_intermediate_results} and
\code{compute_pr_auc_from_curve} for each bootstrap run
}
