% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_helpers.r
\name{generate_replicate_results}
\alias{generate_replicate_results}
\alias{generate_replicate_results_dplyr}
\title{Compute n_bt bootstrap replica, wrapping the composure of
compute_intermediate_results summarise_intermediate_results}
\usage{
generate_replicate_results(
  base_compare,
  n_bt,
  grouping_var,
  seed = NULL,
  ps_flags = list(intermed = FALSE, summarise = FALSE),
  label_distribution = NULL,
  cost_fp = NULL,
  replace_zero_division_with = options::opt("replace_zero_division_with"),
  drop_empty_groups = options::opt("drop_empty_groups"),
  progress = options::opt("progress")
)

generate_replicate_results_dplyr(
  base_compare,
  n_bt,
  grouping_var,
  seed = NULL,
  label_distribution = NULL,
  ps_flags = list(intermed = FALSE, summarise = FALSE),
  cost_fp = NULL,
  progress = FALSE
)
}
\arguments{
\item{base_compare}{a \code{data.frame} as generate by create_comparison}

\item{n_bt}{an integer number of resamples to undergo in bootstrapping}

\item{grouping_var}{character vector of variables that must be present in
base_compare}

\item{seed}{pass seed to resampling step for reproducibility}

\item{ps_flags}{list as returned by \code{set_ps_flags}}

\item{label_distribution}{as in compute_set_retrieval_scores}

\item{cost_fp}{numeric > 0, default is NULL}

\item{replace_zero_division_with}{In macro averaged results (doc-avg, subj-avg) it may occur that some
instances have no predictions or no gold-standard. In these cases calculating
precision and recall may lead to zero divisions. CASIMiR standardly removes
these missing values from macro averages, leading to a smaller \code{support}
(count of instances that were averaged). Other implementations of macro
averaged precision and recall default to 0 in these cases. This option
will allow to control the default. Set any value between 0 and 1. (Defaults to \code{NULL}, overwritable using option 'casimir.replace_zero_division_with' or environment variable 'R_CASIMIR_REPLACE_ZERO_DIVISION_WITH')}

\item{drop_empty_groups}{Should empty levels of factor variables be droped in grouped set retrieval
computation? (Defaults to \code{TRUE}, overwritable using option 'casimir.drop_empty_groups' or environment variable 'R_CASIMIR_DROP_EMPTY_GROUPS')}

\item{progress}{Display progress bars for iterated computations (like bootstrap CI or
PR-Curves) (Defaults to \code{FALSE}, overwritable using option 'casimir.progress' or environment variable 'R_CASIMIR_PROGRESS')}
}
\value{
\code{data.frame} containing \code{n_bt} boot replica of results
as returned by compute_intermediate_results +
summarise_intermediate_results
}
\description{
Compute n_bt bootstrap replica, wrapping the composure of
compute_intermediate_results summarise_intermediate_results
}
\section{Functions}{
\itemize{
\item \code{generate_replicate_results_dplyr()}: variant with dplyr based
internals rather then collapse internals

}}
\examples{

library(casimir)

gold <- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "b",
  "A", "c",
  "B", "a",
  "B", "d",
  "C", "a",
  "C", "b",
  "C", "d",
  "C", "f"
)

pred <- tibble::tribble(
  ~doc_id, ~label_id,
  "A", "a",
  "A", "d",
  "A", "f",
  "B", "a",
  "B", "e",
  "C", "f"
)

base_compare <- casimir:::create_comparison(gold, pred)

boot_replica <- casimir:::generate_replicate_results(
  base_compare,
  n_bt = 10L,
  grouping_var = c("doc_id")
)

}
