% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_helpers.r
\name{helper_f}
\alias{helper_f}
\title{wrapper for compute_intermediate and summarise on one bt_sample}
\usage{
helper_f(
  sampled_id_list,
  compare_cpy,
  grouping_var,
  label_distribution = NULL,
  ps_flags = list(intermed = FALSE, summarise = FALSE),
  cost_fp = NULL
)
}
\arguments{
\item{sampled_id_list}{doc_ids of this bootstrap}

\item{compare_cpy}{as created by \code{create_comparison}}

\item{grouping_var}{variables for aggreation}

\item{label_distribution}{as in compute_set_retrieval_scores}

\item{ps_flags}{list as returned by \code{set_ps_flags}}

\item{cost_fp}{numeric > 0}
}
\value{
as \code{summarise_intermediate_results}
}
\description{
wrapper for compute_intermediate and summarise on one bt_sample
}
