% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_retrieval_scores_reference.r
\name{lrap_score}
\alias{lrap_score}
\title{Helper function for document-wise computation of ranked retrieval scores}
\usage{
lrap_score(gold_vs_pred)
}
\arguments{
\item{gold_vs_pred}{A data.frame as generated by \code{create_comparison}.}
}
\value{
The numeric value of LRAP.
}
\description{
Helper function for document-wise computation of ranked retrieval scores
DCG, NDCG and LRAP.
Implemented as in Annif
\url{https://github.com/NatLibFi/Annif/blob/master/annif/eval.py}.
Reference implementation for Label Ranking Average Precision.
}
