% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ranked-retrieval-scores_helpers.r
\name{ndcg_score}
\alias{ndcg_score}
\title{Reference implementation for ndcg to test against}
\usage{
ndcg_score(gold_vs_pred, limit = NULL)
}
\arguments{
\item{gold_vs_pred}{a \code{data.frame} as generated by create_comparison}

\item{limit}{cut off value for ndcg at N}
}
\value{
numeric value of ndcg
}
\description{
Reference implementation for ndcg to test against
}
