% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_curve_post_processing.r
\name{pr_curve_post_processing}
\alias{pr_curve_post_processing}
\title{Postprocessing of pr curve data}
\usage{
pr_curve_post_processing(results_summary)
}
\arguments{
\item{results_summary}{As produced by \code{summarise_intermediate_results}.}
}
\value{
A data.frame with columns \code{"searchspace_id", "prec", "rec",
  "prec_cummax"} and possible additional stratification variables.
}
\description{
Reshape pr curve data to a format that is easier for plotting.
}
