% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{process_cost_fp}
\alias{process_cost_fp}
\title{Process input for cost_fp}
\usage{
process_cost_fp(cost_fp, gold_vs_pred)
}
\arguments{
\item{cost_fp}{Constant cost assigned to false positives. cost_fp must be
a numeric value > 0 or one of 'max', 'min', 'mean' (computed with reference
to the \code{label_weights})}

\item{gold_vs_pred}{expects \code{data.frame} with cols
\emph{"gold", "label_weight"}.}
}
\value{
a numeric value > 0
}
\description{
Process input for cost_fp
}
