% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-set-retrieval-scores_helpers.r
\name{rename_metrics}
\alias{rename_metrics}
\title{Rename metric names for generalized precision etc.}
\usage{
rename_metrics(res_df)
}
\arguments{
\item{res_df}{data.frame with column \emph{"metric"} containing metric names
prec, rec, f1, rprec}
}
\value{
results data.frame with renamed metrics for generalized precision etc.
}
\description{
Rename metric names for generalized precision etc.
}
