% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{set_grouping_var}
\alias{set_grouping_var}
\title{determine the appropriate grouping variables for each aggregation mode}
\usage{
set_grouping_var(mode, doc_strata, label_dict, var = NULL)
}
\arguments{
\item{mode}{aggregation mode: \emph{"doc-avg", "subj-avg", "micro"}}

\item{doc_strata}{a column that exists gold_standard,
that results should be grouped by, e.g. strata of document-space.
\code{doc_strata} is recommended to be of type factor, so that levels are
not implicitly dropped during bootstrap replications}

\item{label_dict}{two-column \code{data.frame} with col \emph{"label_id"}
and a second column that defines groups of labels to stratify results by.
Results in each stratum
will restrict gold_standard and predictions to the specified label-groups,
as if the vocabulary was consisting of the label group only.
All modes \code{c("doc-avg", "subj-avg", "micro") } are supported within
label-strata.
Nevertheless, mixing \code{mode = "doc-avg"} with fine-grained
label_strata can result in many missing values on document-level results.
Also rank-based thresholding (e.g. Top-5) will result in inhomogeneous
number of labels per documents within the defined label-strata.
\code{mode = "subj-avg"} or \code{mode = "micro"} can be more appropriate
in these circumstances.}

\item{var}{additional variables to include}
}
\value{
a character vector of variables that determine the grouping structure
}
\description{
determine the appropriate grouping variables for each aggregation mode
}
