% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_helpers.R
\name{set_grouping_var}
\alias{set_grouping_var}
\title{determine the appropriate grouping variables for each aggregation mode}
\usage{
set_grouping_var(mode, doc_strata, label_dict, var = NULL)
}
\arguments{
\item{mode}{aggregation mode for later computation. One of c("doc-avg", "subj-avg", "micro")}

\item{doc_strata}{as in compute_set_retrieval_scores}

\item{label_dict}{as in compute_set_retrieval_scores}

\item{var}{additional variables to include}
}
\value{
a character vector of variables that determine the grouping structure
}
\description{
determine the appropriate grouping variables for each aggregation mode
}
