% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_intermediate_results.r
\name{summarise_intermediate_results_dplyr}
\alias{summarise_intermediate_results_dplyr}
\title{Compute mean of intermediate results. Variant with dplyr based
internals rather then collapse internals}
\usage{
summarise_intermediate_results_dplyr(
  intermediate_results,
  propensity_scored = FALSE,
  label_distribution = NULL
)
}
\arguments{
\item{intermediate_results}{as produced by compute intermediate results.
This requires a list containing
\itemize{
\item \code{grouping_var} a  character vector of variables to group by
\item \code{results_table}  a  with cols "prec", "rprec", "rec", "f1" as
}}

\item{propensity_scored}{logical, whether to use propensity scores as weights}

\item{label_distribution}{expects \code{data.frame} with cols
\emph{"label_id", "label_freq", "n_docs"}. \code{label_freq} corresonds to
the number of occurences a label has in the gold_standard. \code{n_docs}
corresponds to the total number of documents in the gold_standard.}
}
\value{
data.frame with cols metric, value
}
\description{
Compute mean of intermediate results. Variant with dplyr based
internals rather then collapse internals
}
